move URL routes and support into a separate file
